LPS-55026 Move to a proper folder
