+++ modules/apps/calendar/calendar-test/test/integration/com/liferay/calendar/service/test/CalendarBookingLocalServiceTest.java
@@ -12,12 +12,13 @@
  * details.
  */
 
-package com.liferay.calendar.service;
+package com.liferay.calendar.service.test;
 
 import com.liferay.calendar.model.Calendar;
 import com.liferay.calendar.model.CalendarBooking;
 import com.liferay.calendar.model.CalendarBookingConstants;
 import com.liferay.calendar.model.CalendarResource;
+import com.liferay.calendar.service.CalendarBookingLocalServiceUtil;
 import com.liferay.calendar.util.CalendarResourceUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
