+++ modules/apps/calendar/calendar-test/test/integration/com/liferay/calendar/upgrade/test/UpgradeCalendarTest.java
